+++ web/lti.go
@@ -15,7 +15,7 @@ func (w *Web) InitLti() {
 
 func loginWithLti(c *Context, w http.ResponseWriter, r *http.Request) {
 	if !c.App.Config().LTISettings.Enable {
-		c.Err =  model.NewAppError("loginWithLti", "api.lti.disabled.app_error", nil, "", http.StatusNotImplemented)
+		c.Err = model.NewAppError("loginWithLti", "api.lti.disabled.app_error", nil, "", http.StatusNotImplemented)
 		return
 	}
 
@@ -25,6 +25,7 @@ func loginWithLti(c *Context, w http.ResponseWriter, r *http.Request) {
 	var ltiConsumerSecret string
 
 	for _, val := range lmss {
+		// TODO: Figure out a better way to find consumer secret for multiple LMSs
 		if lms, ok := val.(model.EdxLMSSettings); ok {
 			if lms.OAuth.ConsumerKey == ltiConsumerKey {
 				ltiConsumerSecret = lms.OAuth.ConsumerSecret
@@ -33,12 +34,17 @@ func loginWithLti(c *Context, w http.ResponseWriter, r *http.Request) {
 		}
 	}
 
+	if ltiConsumerSecret == "" {
+		c.Err = model.NewAppError("loginWithLti", "api.lti.client_secret.app_error", nil, "", http.StatusNotImplemented)
+		return
+	}
+
 	p := utils.NewProvider(ltiConsumerSecret, fmt.Sprintf("%s%s", c.GetSiteURLHeader(), c.Path))
 	p.ConsumerKey = ltiConsumerKey
 	if ok, err := p.IsValid(r); err != nil || ok == false {
 		// TODO: update this based on how we handle request validation error
 		mlog.Error("Invalid LTI request: " + err.Error())
-		c.Err =  model.NewAppError("loginWithLti", "api.lti.validate.app_error", nil, "", http.StatusNotImplemented)
+		c.Err = model.NewAppError("loginWithLti", "api.lti.validate.app_error", nil, "", http.StatusNotImplemented)
 		return
 	}
